test: 1. 컨버터(구분자) 테스트 코드 작성
